fix: refactor seperate tax calculator func
